+++ include/rut/compiler/rir_builder.h
@@ -120,17 +120,17 @@ struct Builder {
         return fn;
     }
 
-    // Record the wait(ms) list for a function: arena-allocates a u16 array
+    // Record the wait(ms) list for a function: arena-allocates a u32 array
     // sized to count and copies the ms values in order. After this call,
     // fn->yield_count reflects the number of state-machine yield points
     // and codegen can consume fn->yield_payload[i].
-    VoidResult set_yield_payload(Function* fn, const u16* ms_list, u32 count) {
+    VoidResult set_yield_payload(Function* fn, const u32* ms_list, u32 count) {
         if (count == 0) {
             fn->yield_count = 0;
             fn->yield_payload = nullptr;
             return {};
         }
-        auto* buf = mod->arena->alloc_array<u16>(count);
+        auto* buf = mod->arena->alloc_array<u32>(count);
         if (!buf) return err(RirError::OutOfMemory);
         for (u32 i = 0; i < count; i++) buf[i] = ms_list[i];
         fn->yield_payload = buf;
